requirements.txt
asgiref==3.7.2
attrs==23.1.0
dj-rest-auth==4.0.1
Django==4.2.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.3
inflection==0.5.1
jsonschema==4.17.3
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
uritemplate==4.1.1
